requirements.txt
absl-py==1.4.0
affine==2.4.0
astunparse==1.6.3
attrs==22.2.0
boto3==1.26.82
botocore==1.29.82
branca==0.6.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Fiona==1.9.1
flatbuffers==23.1.21
folium==0.14.0
fonttools==4.38.0
gast==0.4.0
geopandas==0.12.2
google-auth==2.16.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.0
munch==2.5.0
numpy==1.24.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pyproj==3.4.1
pystac==1.6.1
pystac-client==0.6.0
python-dateutil==2.8.2
pytz==2022.7.1
rasterio==1.3.6
requests==2.28.2
requests-oauthlib==1.3.1
rioxarray==0.13.4
rsa==4.9
s3transfer==0.6.0
shapely==2.0.1
six==1.16.0
snuggs==1.4.7
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
wrapt==1.15.0
xarray==2023.2.0
